requirements.txt
--index-url https://artifacts.jpmchase.net/artifactory/api/pypi/pypi/simple
aiohttp==3.11.10
atlassian-python-api==3.41.19
azure-identity==1.21.0
azure_storage==0.36.0
beautifulsoup4==4.13.4
boto3==1.37.10
faiss_cpu==1.10.0
langchain==0.3.24
langchain_openai==0.3.14
nest-asyncio==1.6.0
#nltk==3.9.1
numpy==2.0.2
openai==1.77.0
Requests==2.32.3
requests-aws4auth==1.3.1
requests_oauthlib==2.0.0
streamlit==1.44.0
tenacity==9.0.0
tiktoken==0.9.0
sumy==0.11.0
scikit-learn
pandas
fuzzywuzzy
cachetools==5.5.2
pytest
levenshtein
python-Levenshtein
azure-core
nltk
rank-bm25
opensearch-py
faiss-cpu
boto3